Apply PR suggestions moving integration tests to be part of go tests
