+++ pkg/cmd/plugincheck2/testdata/grafana-clock-panel-2.1.5.any.zip
