+++ pkg/cmd/plugincheck2/testdata/integration-tests.yaml
