+++ pkg/cmd/plugincheck2/testdata/invalid.zip
