+++ pkg/cmd/plugincheck2/testdata/invalid2.zip
